Tempus-990 Changed size type to double in FileMetaData.

+++ common/data/src/main/java/com/hashmapinc/server/common/data/upload/FileMetaData.java
@@ -20,8 +20,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Date;
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
@@ -29,5 +27,5 @@ public class FileMetaData {
     private String fileName;
     private String extension;
     private long lastUpdated;
-    private long size;
+    private double size;
 }
